requirements.txt
pretty_midi
music21
langchain>=0.1.16
langchain-google-genai>=1.0.4
google-genai
python-dotenv
pydantic
reportlab
